+++ src/integrations/default.ts
@@ -1,4 +1,5 @@
-import { type Integration,consoleLoggingIntegration } from '@sentry/core';
+import { breadcrumbsIntegration, browserApiErrorsIntegration, browserSessionIntegration, globalHandlersIntegration } from '@sentry/browser';
+import { type Integration,dedupeIntegration, functionToStringIntegration, inboundFiltersIntegration, linkedErrorsIntegration } from '@sentry/core';
 import type { CapacitorOptions } from '../options';
 import { deviceContextIntegration } from './devicecontext';
 import { eventOriginIntegration } from './eventorigin';
@@ -18,11 +19,30 @@ export function getDefaultIntegrations(
   integrations.push(nativeReleaseIntegration());
   integrations.push(eventOriginIntegration());
   integrations.push(sdkInfoIntegration());
-  integrations.push(consoleLoggingIntegration());
 
   if (options.enableNative) {
     integrations.push(deviceContextIntegration());
   }
+
+  // @sentry/browser integrations
+  integrations.push(
+    // eslint-disable-next-line deprecation/deprecation
+    inboundFiltersIntegration(),
+    functionToStringIntegration(),
+    browserApiErrorsIntegration(),
+    breadcrumbsIntegration(),
+    globalHandlersIntegration(),
+    linkedErrorsIntegration(),
+    dedupeIntegration(),
+  );
+
+  if (options.enableAutoSessionTracking && !options.enableNative) {
+    integrations.push(browserSessionIntegration());
+  }
+  // end @sentry/browser integrations
+
+  // @sentry/vue integrations must be added manually.
+  // @sentry/react, @sentry/angular @sentry/nuxt dont require any integrations.
 
   return integrations;
 }